+++ R/strip-prefix.R
@@ -1,6 +1,6 @@
 #' Reverse the Q-View internal naming convention
 #'
-#' On import, Q-View Software prefixes well-assignment template names
+#' On import, the producing software prefixes well-assignment template names
 #' with single-letter codes:
 #'
 #' * `"Cal 1"` ... `"Cal N"` -> `"ICal 1"` ... `"ICal N"` (Internal calibrator)